# src/main/java/com/ejemplo/demo/api/dto/ProductoResponse.java
package com.ejemplo.demo.api.dto;

import java.math.BigDecimal;

public record ProductoResponse(
    Long id,
    String nombre,
    BigDecimal precio,
    Long categoriaId,
    String categoriaNombre
) {}